Add comments | Correct light | do some cleaning | Better default value for positions | Add samples [ A

+++ src/jhelp/engine/io/md2/LoaderMD2.java
@@ -863,7 +863,7 @@ public class LoaderMD2
       Triangle triangle;
       Frame frame;
 
-      int index;
+      long index;
 
       while(indexedInputStream.isFinish() == false)
       {
